requirements.txt
langchain-community 
langchain-core 
langchain-google-genai 
pypdf 
chromadb